backend/requirements.txt
boto3==1.34.0
aws-cdk-lib==2.100.0
constructs>=10.0.0